Add coverage no processo de CI
