+++ conf/Conf.php
@@ -10,5 +10,6 @@ final class Conf
     {
         define('URL_HOST', filter_input(INPUT_SERVER, 'HTTP_HOST', FILTER_SANITIZE_URL));
         define('URL', filter_input(INPUT_SERVER, 'HTTP_HOST') . filter_input(INPUT_SERVER, 'REQUEST_URI'));
+        define('PATH_PROJECT', dirname(__DIR__));
     }
 }
